pyproject.toml
[project]
name = "llama-flow"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "llama-index-embeddings-gemini>=0.3.0",
    "llama-index>=0.12.8",
    "llama-index-llms-groq>=0.3.1",
    "llama-parse>=0.5.18",
    "python-dotenv>=1.0.1",
]
